feat(itineraries): itinerary refactor and major cleanup
Abstracted the markup to make this component easier to scan and
understand. Removed the table layout for itineraries because it
mostly sucked and was too tiny. Added a dynamic itinerary display
below the radio inputs instead because that's easier to read and
a better place to display that information anyways.